zeroize: factor integration tests into `tests/` (#771)
